[2/2] OmniGears: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

also cleanup menu settings

Change-Id: I1d6b538f9474f01cbe093d33983e8278fc44c1e8

Conflicts:
	res/xml/menus_settings.xml
	src/org/omnirom/omnigears/interfacesettings/MenusSettings.java